Separate TC string and TC model into separate files.
